move request exceptions to their own namespace
