+++ src/DTS/UseCase/Exception/InvalidGetTemplateRequestException.php
@@ -1,7 +1,7 @@
 <?php declare(strict_types=1);
 
 
-namespace HomeCEU\DTS\UseCase;
+namespace HomeCEU\DTS\UseCase\Exception;
 
 
 class InvalidGetTemplateRequestException extends \Exception {
