+++ src/DTS/UseCase/Exception/Render/InvalidHotRenderRequestException.php
@@ -1,7 +1,7 @@
 <?php declare(strict_types=1);
 
 
-namespace HomeCEU\DTS\UseCase\Render;
+namespace HomeCEU\DTS\UseCase\Exception\Render;
 
 
 class InvalidHotRenderRequestException extends \Exception {
